refactor: move DataHealthChecker to components and remove managers directory
## Rationale
The DataHealthChecker belongs with other data components (GapAnalyzer,
ProgressManager) rather than in a separate managers directory.

## Changes
- Move DataHealthChecker from managers/ to components/
- Update all import references (DataManager, tests)
- Remove now-empty managers/ directory and test directory
- Update components __init__.py to include DataHealthChecker
- Fix import ordering with Ruff

## Verification
- All 1078 unit tests pass
- Quality checks pass (lint, format, typecheck)
- Logical grouping: all data utility classes now in components/
