+++ ktrdr/data/components/__init__.py
@@ -1,6 +1,7 @@
 """Data components module."""
 
+from .data_health_checker import DataHealthChecker
 from .gap_analyzer import GapAnalyzer
 from .progress_manager import ProgressManager, ProgressState
 
-__all__ = ["GapAnalyzer", "ProgressManager", "ProgressState"]
+__all__ = ["DataHealthChecker", "GapAnalyzer", "ProgressManager", "ProgressState"]
