+++ ktrdr/data/data_manager.py
@@ -19,6 +19,7 @@ from ktrdr import (
     log_entry_exit,
     log_performance,
 )
+from ktrdr.data.components.data_health_checker import DataHealthChecker
 from ktrdr.data.components.gap_analyzer import GapAnalyzer
 from ktrdr.data.components.progress_manager import ProgressManager
 from ktrdr.data.data_quality_validator import DataQualityValidator
@@ -27,7 +28,6 @@ from ktrdr.data.gap_classifier import GapClassifier
 from ktrdr.data.ib_data_adapter import IbDataAdapter
 from ktrdr.data.loading_modes import DataLoadingMode
 from ktrdr.data.local_data_loader import LocalDataLoader
-from ktrdr.data.managers.data_health_checker import DataHealthChecker
 from ktrdr.data.timeframe_synchronizer import TimeframeSynchronizer
 from ktrdr.errors import (
     DataCorruptionError,
