--- ktrdr/data/managers/__init__.py
@@ -1,7 +1,0 @@
-"""
-Data Manager modules for ktrdr data management system.
-
-This package contains manager utilities for data operations.
-"""
-
-__all__: list[str] = []
